chore: Re-organise examples
