+++ examples/webhook/heroku.js
@@ -5,7 +5,7 @@
 
 
 const TOKEN = process.env.TELEGRAM_TOKEN || 'YOUR_TELEGRAM_BOT_TOKEN';
-const TelegramBot = require('..');
+const TelegramBot = require('../..');
 const options = {
   webHook: {
     // Port to which you should bind is assigned to $PORT variable
