+++ examples/webhook/https.js
@@ -5,12 +5,12 @@
 
 
 const TOKEN = process.env.TELEGRAM_TOKEN || 'YOUR_TELEGRAM_BOT_TOKEN';
-const TelegramBot = require('..');
+const TelegramBot = require('../..');
 const options = {
   webHook: {
     port: 443,
-    key: `${__dirname}/key.pem`,  // Path to file with PEM private key
-    cert: `${__dirname}/crt.pem`  // Path to file with PEM certificate
+    key: `${__dirname}/../ssl/key.pem`,  // Path to file with PEM private key
+    cert: `${__dirname}/../ssl/crt.pem`  // Path to file with PEM certificate
   }
 };
 // This URL must route to the port set above (i.e. 443)
