+++ examples/webhook/now.js
@@ -6,7 +6,7 @@
 
 
 const TOKEN = process.env.TELEGRAM_TOKEN || 'YOUR_TELEGRAM_BOT_TOKEN';
-const TelegramBot = require('..');
+const TelegramBot = require('../..');
 const options = {
   webHook: {
     // Just use 443 directly
